+++ src/components/keyboard/MahjongKeyboard.css
