+++ src/content/pointPredicateContent.tsx
