+++ src/content/pointPredicateIdToContentMap.tsx
